Update encoder library
Based on keyboards/nightly_boards/\*/encoder_action.\* by Neil Brian Ramirez, new method for setting encoder actions. This method requires that the actions are defined in the keymap. Switch Ortho5x12 to use the handwired/5x13 layout, with the first column being a dummy that will set the encoder actions.

(WIP) Also started working on an alternate version that will support keymaps for existing keyboards that do not have extra places in the layout to add the encoder keys (e.g., the Gherkin). This will use an additional layer that will define all the encoder actions for the other layers based on the position.